Removing pagination model
As a model isn't good, tried traits but no good either
can't get genric enough for a trait and also needed to send SQL to limit query. Not good in a controller which shouldn't know anything about the tables.

Just going for a FAT model.
