+++ App/Controllers/Category.php
@@ -3,7 +3,6 @@
 namespace App\Controllers;
 
 use App\Models\CategoryModel;
-use App\Models\PaginationModel;
 use App\Models\PostModel;
 use App\Models\SlugModel;
 use Core\Container;
@@ -18,7 +17,6 @@ class Category extends Controller
     private $slugModel;
     private $postModel;
     private $categoryModel;
-    private $paginationModel;
 
     public function __construct(Container $container)
     {
@@ -28,7 +26,6 @@ class Category extends Controller
         $this->slugModel = new SlugModel($this->container);
         $this->postModel = new PostModel($this->container);
         $this->categoryModel = new CategoryModel($this->container);
-        $this->paginationModel = new PaginationModel($this->container);
 
         $this->sendSessionVars();
 
@@ -49,7 +46,7 @@ class Category extends Controller
 
 
         $categoryId = $this->slugModel->getIdFromSlug($categorySlug, "categories", "categories_slug", "idcategories");
-        $totalPosts = $this->paginationModel->totalNumberPostsInCategory($categoryId);
+        $totalPosts = $this->postModel->totalNumberPostsInCategory($categoryId);
 
         $pagination = $this->pagination->getPagination($page, $totalPosts);
 
@@ -76,7 +73,7 @@ class Category extends Controller
      */
     public function allPosts(string $page = "page-1")
     {
-        $totalPosts = $this->paginationModel->totalNumberPosts();
+        $totalPosts = $this->postModel->totalNumberPosts();
 
         $pagination = $this->pagination->getPagination($page, $totalPosts);
